--- components/eamxx/src/python/pyeamxx/__init__.py
@@ -1,23 +1,0 @@
-"""
-    This file will serve as a way to organize and expose
-    libpyscream internals to the rest of pyscream
-"""
-
-from libpyscream.libpyscream_ext import AtmProc
-from libpyscream.libpyscream_ext import Grid
-from libpyscream.libpyscream_ext import ParameterList
-from libpyscream.libpyscream_ext import init
-from libpyscream.libpyscream_ext import Field
-from libpyscream.libpyscream_ext import P3
-from libpyscream.libpyscream_ext import finalize
-
-
-__all__ = [
-    'init',
-    'finalize',
-    'AtmProc',
-    'Grid',
-    'ParameterList',
-    'Field',
-    'P3',
-]
