+++ components/eamxx/src/python/pyscream/__init__.py
@@ -1,0 +1,10 @@
+"""
+    This file will serve as a way to organize and expose
+    libpyscream internals to the rest of pyscream
+"""
+
+from libpyscream import pyscream_ext as pyscream
+
+__all__ = [
+    "pyscream"
+]
